+++ src/font_override.hpp
@@ -11,6 +11,8 @@ void initialize_font_override();
 void shutdown_font_override();
 void begin_item_prompt_font();
 void end_item_prompt_font();
+void begin_map_font();
+void end_map_font();
 bool draw_font_override(void* args, void* retval, FontDrawOriginal drawOriginal);
 
 }  // namespace twilight_hd_hud